Show skin preview in skin list
In order to do this, we migrated the cell-based table view to a view-based table view.

Also, we edited the storyboard by hand to add the binding we want for the SkinView view…
Let's hope this manual edition stands! We did not edit the storyboard with the GUI before the
commit to be sure to have something that works; we may have to bind in code instead later…